+++ aim_flash_storage/examples/stm32_flash/src/main.cpp
@@ -25,8 +25,9 @@ static uint32_t s_rowCounter = 0U;
 // --- Application hook ---
 
 static void handleStatus(Stream& out) {
-  out.printf("uptime=%lu ms  rows=%u\r\n",
-             static_cast<unsigned long>(millis()), s_rowCounter);
+  out.printf("uptime=%lu ms  rows=%lu\r\n",
+             static_cast<unsigned long>(millis()),
+             static_cast<unsigned long>(s_rowCounter));
 }
 
 static const AimConsoleHook kHooks[] = {
